+++ src/lib.rs
@@ -60,7 +60,7 @@ pub use adapter::{
 };
 pub use amf::Amf0Value;
 pub use amf3::Amf3Value;
-pub use client::{RtmpClient, RtmpUrl};
+pub use client::{ClientEvent, RtmpClient, RtmpUrl};
 pub use error::{Error, Result};
 pub use flv::{AudioTag, ModEx, VideoTag};
 pub use server::{PublishRequest, RtmpServer, RtmpSession, StreamPacket};